[CON-946] Behave as if connected to power if no battery present
Pegasus doesn't correctly report usb power connected status on Pegasus.
It does, however, correctly report that the battery is absent.  When we
detect that the battery is absent, report to mtab/cts that the battery
is at 100% and we are connected to power (the latter of which, at least,
must be true if there is no battery).

[CON-946](https://fivestars.atlassian.net/browse/CON-946)